Stop throwing an exception for a situation that is not an error
It is possible for different instances of an application to make different
choices about whether to cache an entity or not. If one instance is caching
it, it correctly sends updates out over the message queue to other
instances. But if a non-caching instance receives such a message, it's not
actually an error. Ignore instead of throwing an exception.