Take a stab at implementing AnnotationDispatcher
This is MOSTLY copy-paste of BasicDispatcher, PathDispatcher, and
MethodUriHandler as a proof-of-concept. The idea is that an application
could override `constructDispatcher` and instantiate a new
`AnnotationDispatcher`, and that would use reflection to find classes
annotated with `@Path`, create `AnnotationHandler` instances for those
classes with a reference to that class, and it would do most of what
MethodUriHandler does once it has found a subclass of MethodUriHandler -
that is, inspect its own methods (however, it is changes to the
reference class) looking for `@Path` and `@GET/etc` annotations and
building routing trees for that reference class.

The theory then is that a simple `constructDispatcher` implementation
would allow the following POJO to be a handler:

@Path
public class RootHandler {
  @Path
  @GET
  public String handleRoot() {
    return "";
  }
}

Note: the return type of the method is still WIP and will eventually
allow methods to be annotated with 'producer' annotations to specify the
strategy for serializing the return type. Something like `@ProducesJson`
(not yet finalized) would indicate to the dispatcher that that `String`
return type should be turned into JSON (how? Not sure yet, but for more
complicated POJO types, Jackson can serialize fairly easily).