Improve performance of getObjects() for EntityGroup
MethodValueCache's getObjects() implementation was fine for a CacheGroup
because everything was already in memory and fast to fetch. However, an
EntityGroup would end up doing a separate query for every requested entity
and this could be quite slow.

This change simply uses the existing functionality of map() to get all
requested entities in a single query so it is faster for an EntityGroup.
Performance for CacheGroups is not impacted significantly.