When invalidating the session make it possible to start using a new one
Fix an issue where logout would correctly invalidate the session, but the
Context.session remained the invalidated one so any subsequent changes to
session values during that response would be lost. This change clears
Context.session so a new session is created as needed and available for use
during this response, instead of needing to wait for a subsequent request
for the new session to be available.